% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-data-descr.R
\name{dl_data_descr}
\alias{dl_data_descr}
\title{Download data description (README)}
\usage{
dl_data_descr(overwrite = FALSE)
}
\arguments{
\item{overwrite}{a logical. If \code{TRUE}, the file will be downloaded again and
the previous version will be erased.}
}
\value{
No return value.
}
\description{
This function downloads the \code{README.md} of the GitHub repository
\url{https://github.com/rdatatoolbox/datarepo/} providing data description.
The file won't be downloaded if already exists locally (except if
\code{overwrite = TRUE}).
The \code{README.md} will be stored in \verb{data/}.
}
\examples{
\dontrun{
dl_data_descr()
}
}
