% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-wildfinder-data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Download WWF WildFinder datasets}
\usage{
dl_wildfinder_data(overwrite = FALSE)
}
\arguments{
\item{overwrite}{a logical. If \code{TRUE}, the files will be downloaded again
and the previous versions will be erased.}
}
\value{
No return value.
}
\description{
This function downloads the WWF WildFinder dataset (csv files) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}. The files
won't be downloaded if already exist locally (except if \code{overwrite = TRUE}).
Three \code{.csv} files will be stored in \verb{data/wildfinder/}:
\itemize{
\item \code{wildfinder-ecoregions_list.csv}
\item \code{wildfinder-ecoregions_species.csv}
\item \code{wildfinder-mammals_list.csv}
}

The folder \verb{data/wildfinder/} will be created if required.
}
\examples{
\dontrun{
dl_wildfinder_data()
}
}
